refactor: inline submenu mixin, add button and submenu typings
