+++ packages/menu-bar/src/vaadin-menu-bar-button.js
@@ -54,3 +54,5 @@ class MenuBarButton extends Button {
 }
 
 defineCustomElement(MenuBarButton);
+
+export { MenuBarButton };
